refactoring extension stuff to improve pebble engine initialization
much of this refactoring is built around the idea of absolutely NOT
wanting applications using this library to be required to explicitly
call some kind of "init" function before templates can be rendered.

however, for this library's (eventual) unit tests, we will want the
ability to reset the engine state (reset-engine! function), so a way
to add extensions to an arbitrary PebbleEngine object is needed